Revised code list
Removed the CPT code "82105" from the PRENATAL codelist. This corresponded to a serum-based alpha-fetal protein test. While typically only used in pregnancy, we learned that this test is also used to diagnosis/monitor specific cancers and liver conditions. As a result, we decided it was inappropriate to identify pregnancies with this code if no other code was present. We kept it in the GEST_AGE sheet and additionally added the CPT code "82106" to the GEST_AGE sheet because it also corresponds to the alpha-fetal protein test but uses amniotic fluid instead of blood serum.